requirements.txt
# Requirements for docker-compose generation.
PyYAML~=6.0.2